SQL verb fixes

+++ pricedb/db/schema/prices/20250117004126_init.sql
@@ -1,7 +1,7 @@
 -- migrate:up
 CREATE TABLE prices (
     code VARCHAR(10) NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
+    price DOUBLE NOT NULL,
     currency VARCHAR(3) NOT NULL,
     timestamp TIMESTAMP NOT NULL
 );
